fix(dashboard): only show bundle sync callout when attention needed
